requirements.txt
aiohttp==3.9.5
twitchio==2.9.1
google-generativeai==0.6.0
